% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petit_plot.R
\name{petit_plot}
\alias{petit_plot}
\alias{petit_ggplot}
\title{Save a plot to a file and automatically shrink it}
\usage{
petit_plot(
  filename = "plot",
  path = NULL,
  device = "png",
  ragg = FALSE,
  keep_large = FALSE,
  suffix,
  quiet,
  return_path,
  key = NULL,
  ...
)

petit_ggplot(
  filename = "plot",
  path = NULL,
  plot = ggplot2::last_plot(),
  device = "png",
  keep_large = FALSE,
  suffix,
  quiet,
  return_path,
  key = NULL,
  ...
)
}
\arguments{
\item{filename}{String, required. The name to give the output image file. Do
not include a file extension.}

\item{path}{String, optional. If \code{NULL}, defaults to the current working directory. The
path to save the image file into, relative to the current working directly. Do not include
the final trailing path separator.}

\item{device}{String, optional. Defaults to \code{"png"}. One of \code{"png"} or \code{"jpg"},
to choose the output image file type. TinyPNG only supports \code{png} or \code{jpg} image types.}

\item{ragg}{Boolean, optional. Defaults to \code{FALSE}. Whether to use the \href{https://ragg.r-lib.org}{ragg}
package as the device backend to generate the image. Will error if \href{https://ragg.r-lib.org}{ragg} is
not installed, but does not need to be loaded.}

\item{keep_large}{Boolean, optional. Defaults to \code{FALSE}. Whether to keep the
unshrunk original image file alongside the tiny version, or just keep the
shrunk file.}

\item{suffix}{String, optional. If \code{keep_large} is \code{TRUE}, the suffix to add
to the shrunk file. Ignored if \code{keep_large} is \code{FALSE}.}

\item{quiet}{Boolean, optional. If set to \code{TRUE}, \code{tinify()} displays no
information messages as it shrinks files.}

\item{return_path}{String or \code{NULL}, optional. One of "\code{proj}", "\code{rel}", "\code{abs}", or
"\code{all}". If "\code{proj}", will return the file path of the newly tinified image
file relative to the Rstudio project directory (looking for an .Rproj file).
If no project can be identified, returns \code{NA}. If "\code{rel}", will return the
file path of the newly tinified image file, relative to the \strong{current}
working directory at the time \code{tinify()} is called. If "\code{abs}", will return
the absolute file path of the newly tinified image file. If "\code{all}", will
return a named list with all file paths. If \code{NULL} (the default), no file
path is returned.}

\item{key}{String, optional. A string containing your TinyPNG API key. Not
required if the API key is set using \code{tinify_api()}. Any other key provided
as an argument will override the key set by \code{tinify_api()}.}

\item{...}{Additional plot options, passed directly to either \code{png()}/\code{ragg::agg_png()},
\code{jpeg()}/\code{ragg::agg_jpeg()}, or \code{ggplot2::ggsave()} depending on method used.}

\item{plot}{Object, optional. The plot object to export. Defaults to the last plot
modified or created (using \code{ggplot2::last_plot()}) if not provided.}
}
\value{
If \code{return_path = "proj"}, \code{return_path = "rel"}, or \code{return_path = "abs"}, a string with the project, relative, or absolute path to the newly
tinified image file. If no project can be identified for \code{return_path = "proj"}, returns \code{NA}. If \code{return_path = "all"}, a named list with all file
paths included as \verb{$project}, \verb{$relative}, and \verb{$absolute} respectively. If
\code{return_path = NULL}, no return value.
}
\description{
Save a plot to a file and automatically shrink it with \code{tinify()}.
}
\details{
These are convenience functions to wrap saving either a base R plot with
\code{png()} or \code{jpeg()} devices, or a \href{https://ggplot2.tidyverse.org}{ggplot} object with \code{ggplot2::ggsave()}, before passing the resulting
image file directly to \code{tinify()}.  Can also make use of the \href{https://ragg.r-lib.org}{ragg}
package for base plots if installed.
}
\section{Using with base R plots}{
 To save and shrink a base R plot, print the plot and call
\code{petit_plot()} immediately after. Under the hood, this uses \code{recordPlot()} to capture
and replay the last plot created within the chosen device with the applied options:\preformatted{plot(mtcars$mpg, mtcars$drat)

petit_plot(filename = "mtcars")
}
}

\section{Using with ggplot}{
 To save and shrink a \href{https://ggplot2.tidyverse.org}{ggplot}, either create, modify, or show
the plot and call \code{petit_plot()} immediately after, in a similar process to base R plots:\preformatted{ggplot(data = palmerpenguins::penguins,
       aes(flipper_length_mm, body_mass_g)) +
geom_point(aes(color = species))

petit_plot(filename = "penguins")
}

Or use \code{petit_ggplot()} to capture specifically the last ggplot created or modified:\preformatted{ggplot(data = palmerpenguins::penguins,
       aes(flipper_length_mm, body_mass_g)) +
geom_point(aes(color = species))

petit_ggplot(filename = "penguins")
}

Or provide the plot object explicitly to \code{petit_ggplot()} with \code{plot}:\preformatted{p <- ggplot(data = palmerpenguins::penguins,
            aes(flipper_length_mm, body_mass_g)) +
     geom_point(aes(color = species))

petit_ggplot(filename = "penguins", plot = p)
}
}

\seealso{
\code{\link[=tinify]{tinify()}}

\code{\link[=tinify_key]{tinify_key()}} to set an API key globally

\code{\link[=tinify_defaults]{tinify_defaults()}} to set default arguments that will be used if not
provided explicitly
}
