% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinify_defaults.R
\name{tinify_defaults}
\alias{tinify_defaults}
\title{Set defaults for \code{tinify()} function}
\usage{
tinify_defaults(overwrite, suffix, quiet, return_path, resize)
}
\arguments{
\item{overwrite}{Boolean, defaults to \code{FALSE}. By default, tinify will create
a new file with the suffix '_tiny' and preserve the original file. Set
\code{TRUE} to instead overwrite the original file, with the same file name.}

\item{suffix}{String, defaults to \code{"_tiny"}. By default, tinify will create a
new file with the suffix '_tiny' and preserve the original file. Provide a
new character string here to change the suffix from '_tiny' to your own
choice. Empty strings (\code{""}) are not accepted. \code{suffix} is ignored when
\code{overwrite} is set to \code{TRUE}.}

\item{quiet}{Boolean, defaults to \code{FALSE}. By default, tinify provides details
on file names, amount of file size reduction (\% and Kb), and the number of
TinyPNG API calls made this month. If set to \code{TRUE}, tinify displays no
messages as it shrinks files.}

\item{return_path}{String or \code{NULL}, optional. One of "\code{proj}", "\code{rel}", "\code{abs}", or
"\code{all}". If "\code{proj}", will return the file path of the newly tinified image
file relative to the Rstudio project directory (looking for an .Rproj file).
If no project can be identified, returns \code{NA}. If "\code{rel}", will return the
file path of the newly tinified image file, relative to the \strong{current}
working directory at the time \code{tinify()} is called. If "\code{abs}", will return
the absolute file path of the newly tinified image file. If "\code{all}", will
return a named list with all file paths. Set to \code{NULL} to stop returning a
file path.}

\item{resize}{Named list or \code{NULL}, optional. A named list with the elements \code{method} as
a string, and \code{width} and/or \code{height} as numerics. Please note you can only
reduce an image's dimensions and make an image smaller with TinyPNG API, not
make an image larger. Method must be set to one of "scale", "fit", "cover",
or "thumb". If using "scale", you only need to provide \code{width} OR \code{height},
not both. If using any other method, you must supply both a \code{width} AND
\code{height}. See \url{https://tinypng.com/developers/reference#resizing-images} and
the examples for more. Set to \code{NULL} to stop resizing.}
}
\description{
Set some default options for \code{tinify()} in the global environment, so it is no
longer necessary to explicitly provide each argument with every call of
\code{tinify()}.

If called without any arguments, \code{tinify_defaults()} will print
the current default options set to the console.
}
\examples{
\dontrun{

tinify_defaults(quiet = TRUE, suffix = "_small")

# show current defaults set

tinify_defaults()
#> Tinify 'overwrite' set to: FALSE
#> Tinify 'suffix' set to: "_small"
#> Tinify 'quiet' set to: FALSE
#> Tinify 'return_path' set to: No return
#> Tinify 'resize' set to: No resize

}
}
\seealso{
\code{\link[=tinify]{tinify()}} to shrink image filesizes

\code{\link[=tinify_key]{tinify_key()}} to set a default TinyPNG.com API key
}
