% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinify_key.R
\name{tinify_key}
\alias{tinify_key}
\title{Set TinyPNG.com API Key}
\usage{
tinify_key(key)
}
\arguments{
\item{key}{A string containing your TinyPNG.com API key.}
}
\description{
Set your TinyPNG API key in the system environment, so it is
no longer necessary to explicitly provide an API key with every
call of \code{tinify()}.
}
\details{
You can get a TinyPNG API key from \url{https://tinypng.com/developers}.
}
\examples{
\dontrun{
tinify_key("YOUR-API-KEY-HERE")
}
}
\seealso{
\code{\link[=tinify]{tinify()}} to shrink image filesizes
}
