% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinieR-package.R
\docType{package}
\name{tinieR-package}
\alias{tinieR}
\alias{tinieR-package}
\title{tinieR: Shrink Image Filesizes with TinyPNG.com}
\description{
Shrink image filesizes with the TinyPNG API <https://tinypng.com>. Works with .png and .jpg/.jpeg files, and can return the new image filepath to enable embedding in other image workflows/functions.
}
\details{
Shrink image filesizes with the TinyPNG.com API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jmablog/tinieR}
  \item Report bugs at \url{https://github.com/jmablog/tinieR/issues}
}

}
\author{
\strong{Maintainer}: James Adams \email{james@jmablog.com}

}
\keyword{internal}
